Move PolygonToolForm into its own component.
