+++ frontend/src/features/map_planning/layers/base/components/BaseLayerAttributeEditForm.tsx
@@ -9,9 +9,6 @@ import { useDebouncedSubmit } from '@/hooks/useDebouncedSubmit';
 import { ReactComponent as CheckIcon } from '@/svg/icons/check.svg';
 import { ReactComponent as CircleDottedIcon } from '@/svg/icons/circle-dotted.svg';
 import { ReactComponent as CloseIcon, ReactComponent as CrossIcon } from '@/svg/icons/close.svg';
-import { ReactComponent as EraserIcon } from '@/svg/icons/eraser.svg';
-import { ReactComponent as PencilPlusIcon } from '@/svg/icons/pencil-plus.svg';
-import { ReactComponent as PointerIcon } from '@/svg/icons/pointer.svg';
 import { zodResolver } from '@hookform/resolvers/zod';
 import { useState } from 'react';
 import { useForm } from 'react-hook-form';
@@ -43,14 +40,6 @@ export interface BaseLayerEditFormProps {
 
 export function BaseLayerAttributeEditForm({ onChange, isReadOnlyMode }: BaseLayerEditFormProps) {
   const { t } = useTranslation(['common', 'baseLayerForm']);
-
-  const activatePolygonAddPoints = useMapStore((state) => state.baseLayerActivateAddPolygonPoints);
-  const activatePolygonMovePoints = useMapStore(
-    (state) => state.baseLayerActivateMovePolygonPoints,
-  );
-  const activatePolygonDeletePoints = useMapStore(
-    (state) => state.baseLayerActivateDeletePolygonPoints,
-  );
 
   const activateMeasurement = useMapStore((state) => state.baseLayerActivateMeasurement);
   const deactivateMeasurement = useMapStore((state) => state.baseLayerDeactivateMeasurement);
@@ -94,42 +83,6 @@ export function BaseLayerAttributeEditForm({ onChange, isReadOnlyMode }: BaseLay
 
   return (
     <div className="flex flex-col gap-4">
-      <h2>{t('baseLayerForm:polygon_tools_title')}</h2>
-      <div className="flex flex-row gap-1">
-        <IconButton
-          isToolboxIcon={true}
-          onClick={() => {
-            activatePolygonMovePoints();
-            setStatusPanelContent(
-              <PolygonStatusPanelContent text={t('baseLayerForm:polygon_move_points_hint')} />,
-            );
-          }}
-        >
-          <PointerIcon></PointerIcon>
-        </IconButton>
-        <IconButton
-          isToolboxIcon={true}
-          onClick={() => {
-            activatePolygonAddPoints();
-            setStatusPanelContent(
-              <PolygonStatusPanelContent text={t('baseLayerForm:polygon_add_points_hint')} />,
-            );
-          }}
-        >
-          <PencilPlusIcon></PencilPlusIcon>
-        </IconButton>
-        <IconButton
-          isToolboxIcon={true}
-          onClick={() => {
-            activatePolygonDeletePoints();
-            setStatusPanelContent(
-              <PolygonStatusPanelContent text={t('baseLayerForm:polygon_delete_points_hint')} />,
-            );
-          }}
-        >
-          <EraserIcon></EraserIcon>
-        </IconButton>
-      </div>
       <h2>{t('baseLayerForm:title')}</h2>
       <FileSelectorModal
         setShow={function (show: boolean): void {
@@ -264,30 +217,5 @@ export function BaseLayerAttributeEditForm({ onChange, isReadOnlyMode }: BaseLay
         </div>
       </div>
     </div>
-  );
-}
-
-function PolygonStatusPanelContent(props: { text: string }) {
-  const deactivatePolygonManipulation = useMapStore(
-    (state) => state.baseLayerDeactivatePolygonManipulation,
-  );
-  const clearStatusPanelContent = useMapStore((state) => state.clearStatusPanelContent);
-
-  return (
-    <>
-      <div className="flex flex-row items-center justify-center">{props.text}</div>
-      <div className="flex items-center justify-center">
-        <IconButton
-          className="m-2 h-8 w-8 border border-neutral-500 p-1"
-          onClick={() => {
-            deactivatePolygonManipulation();
-            clearStatusPanelContent();
-          }}
-          data-tourid="placement_cancel"
-        >
-          <CloseIcon></CloseIcon>
-        </IconButton>
-      </div>
-    </>
   );
 }
