+++ frontend/src/features/map_planning/layers/base/components/BaseLayerRightToolbar.tsx
@@ -8,6 +8,7 @@ import {
   BaseLayerDistanceModalAttributes,
   DistanceMeasurementModal,
 } from '@/features/map_planning/layers/base/components/DistanceMeasurementModal';
+import { PolygonToolForm } from '@/features/map_planning/layers/base/components/PolygonToolForm';
 import { calculateDistance, calculateScale } from '@/features/map_planning/layers/base/util';
 import useMapStore from '@/features/map_planning/store/MapStore';
 import { useIsReadOnlyMode } from '@/features/map_planning/utils/ReadOnlyModeContext';
@@ -115,6 +116,7 @@ export const BaseLayerRightToolbar = () => {
         onCancel={deactivateMeasurement}
         show={measureStep === 'both selected'}
       />
+      <PolygonToolForm />
       <BaseLayerAttributeEditForm
         // remount the form when the selected planting or the step changes (on undo/redo)
         key={`${baseLayerState.id}-${step}`}
